Refatoração de código, implementando todos os metodos que a api disponibiliza, arrumando o metodo put, CRUD funcionando AAAAAAAAAAEEEEEEEEEEEEEEEE
