--- Game.App/Controllers/PutController.cs
@@ -1,33 +1,0 @@
-﻿using Game.App.Models;
-using Microsoft.AspNetCore.Mvc;
-using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net.Http;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Game.App.Controllers
-{
-    public class PutController : Controller
-    {
-        [HttpGet]
-        public IActionResult Index()
-        {
-            return View();
-        }
-
-        [HttpPost]
-        public async Task<IActionResult> Update(Guid id, [FromForm] GameModel game)
-        {
-            using (var client = new HttpClient())
-            {
-                var serializedProduto = JsonConvert.SerializeObject(game);
-                var content = new StringContent(serializedProduto, Encoding.UTF8, "application/json");
-                var result = await client.PutAsync($"http://localhost:5000/api/game/{id}", content);
-            }
-            return RedirectToAction("Index", "Home");
-        }
-    }
-}
