+++ Game.App/Properties/serviceDependencies.json
@@ -1,0 +1,8 @@
+{
+  "dependencies": {
+    "mssql1": {
+      "type": "mssql",
+      "connectionId": "GameAppContext"
+    }
+  }
+}
