+++ Game.App/Properties/serviceDependencies.local.json
@@ -1,0 +1,8 @@
+{
+  "dependencies": {
+    "mssql1": {
+      "type": "mssql.local",
+      "connectionId": "GameAppContext"
+    }
+  }
+}
