+++ Game.App/Startup.cs
@@ -1,8 +1,10 @@
-using Microsoft.AspNetCore.Builder;
+﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
+using Microsoft.EntityFrameworkCore;
+using Game.App.Data;
 
 namespace Game.App
 {
@@ -19,6 +21,9 @@ namespace Game.App
         {
             services.AddHttpClient();
             services.AddControllersWithViews();
+
+            services.AddDbContext<GameAppContext>(options =>
+                    options.UseSqlServer(Configuration.GetConnectionString("GameAppContext")));
         }
 
         public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
