+++ Game.App/Views/Delete/Index.cshtml
@@ -1,0 +1,48 @@
+﻿@model Game.App.Models.GameModel
+
+
+<h1>Delete</h1>
+
+<h3>Você quer mesmo apagar este jogo?</h3>
+<div>
+    <h4>GameModel</h4>
+    <hr />
+    <dl class="row">
+        <dt class="col-sm-2">
+            @Html.DisplayNameFor(model => model.Nome)
+        </dt>
+        <dd class="col-sm-10">
+            @Html.DisplayFor(model => model.Nome)
+        </dd>
+        <dt class="col-sm-2">
+            @Html.DisplayNameFor(model => model.Descricao)
+        </dt>
+        <dd class="col-sm-10">
+            @Html.DisplayFor(model => model.Descricao)
+        </dd>
+        <dt class="col-sm-2">
+            @Html.DisplayNameFor(model => model.Multiplayer)
+        </dt>
+        <dd class="col-sm-10">
+            @Html.DisplayFor(model => model.Multiplayer)
+        </dd>
+        <dt class="col-sm-2">
+            @Html.DisplayNameFor(model => model.MaiorDeIdade)
+        </dt>
+        <dd class="col-sm-10">
+            @Html.DisplayFor(model => model.MaiorDeIdade)
+        </dd>
+        <dt class="col-sm-2">
+            @Html.DisplayNameFor(model => model.DataLancamento)
+        </dt>
+        <dd class="col-sm-10">
+            @Html.DisplayFor(model => model.DataLancamento)
+        </dd>
+    </dl>
+
+    <form asp-controller="Delete" asp-action="Index">
+        <input type="hidden" asp-for="GameId" />
+        <input type="submit" value="Delete" class="btn btn-danger" /> |
+        <a asp-action="Index">Back to List</a>
+    </form>
+</div>
