+++ Game.App/Views/Edit/Index.cshtml
@@ -1,19 +1,15 @@
 ﻿@model Game.App.Models.GameModel
 
 
-<h1>Index</h1>
+<h1>Edit</h1>
 
 <h4>GameModel</h4>
 <hr />
 <div class="row">
     <div class="col-md-4">
-        <form asp-controller="Put" asp-action="Update">
+        <form asp-controller="Edit" asp-action="Index">
             <div asp-validation-summary="ModelOnly" class="text-danger"></div>
-            <div class="form-group">
-                <label asp-for="GameId" class="control-label"></label>
-                <input asp-for="GameId" class="form-control" />
-                <span asp-validation-for="GameId" class="text-danger"></span>
-            </div>
+            <input type="hidden" asp-for="GameId" />
             <div class="form-group">
                 <label asp-for="Nome" class="control-label"></label>
                 <input asp-for="Nome" class="form-control" />
@@ -49,3 +45,7 @@
 <div>
     <a asp-action="Index">Back to List</a>
 </div>
+
+@section Scripts {
+    @{await Html.RenderPartialAsync("_ValidationScriptsPartial");}
+}
