+++ Game.App/appsettings.json
@@ -1,6 +1,7 @@
 {
   "ConnectionStrings": {
-    "DefaultConnection": "Server=(localdb)\\mssqllocaldb;Database=Game;Trusted_Connection=True;"
+    "DefaultConnection": "Server=(localdb)\\mssqllocaldb;Database=Game;Trusted_Connection=True;",
+    "GameAppContext": "Server=(localdb)\\mssqllocaldb;Database=GameAppContext-873caed3-596e-40e5-bf77-e69e43e52c69;Trusted_Connection=True;MultipleActiveResultSets=true"
   },
   "Logging": {
     "LogLevel": {
