+++ Games.API/Repositories/GameRepository.cs
@@ -1,5 +1,6 @@
 ﻿using Game.API.Data;
 using Game.API.Models;
+using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Linq;
